requirements.txt
streamlit==1.26.0
mysql-connector-python==8.1.0
pandas==2.0.3
Pillow==10.0.1
opencv-python==4.8.1
torch==2.0.1
facenet-pytorch==2.5.3
torchvision==0.15.2
scikit-learn==1.3.0
numpy==1.24.4